兼容屏幕，未测试
